move ExtWeights to core-processor
